rename decorations files
