+++ main.js
@@ -1,5 +1,5 @@
 import 'reveal.js/dist/reveal.css'
-import './reveal-nlesc.scss'
+import './nlesc-decorations.scss'
 import Reveal from 'reveal.js'
 import Highlight from 'reveal.js/plugin/highlight/highlight.esm.js';
 import Markdown from 'reveal.js/plugin/markdown/markdown.esm.js';
@@ -8,7 +8,7 @@ import Notes from 'reveal.js/plugin/notes/notes.esm.js';
 import Search from 'reveal.js/plugin/search/search.esm.js';
 import Zoom from 'reveal.js/plugin/zoom/zoom.esm.js';
 import { Reveald3 } from './plugin/reveald3.js';
-import Decorations from './decorations.js';
+import Decorations from './nlesc-decorations.js';
 
 const deck = new Reveal()
 deck.initialize({
